+++ lib/streamlit/__init__.py
@@ -83,8 +83,8 @@ from streamlit.proto import ForwardMsg_pb2 as _ForwardMsg_pb2
 
 from streamlit.echo import echo as echo
 from streamlit.legacy_caching import cache as cache
-from streamlit.caching import singleton as experimental_singleton
-from streamlit.caching import memo as experimental_memo
+from streamlit.runtime.caching import singleton as experimental_singleton
+from streamlit.runtime.caching import memo as experimental_memo
 
 # This is set to True inside cli._main_run(), and is False otherwise.
 # If False, we should assume that DeltaGenerator functions are effectively
@@ -406,7 +406,7 @@ def spinner(text: str = "In progress...") -> Iterator[None]:
 
     """
     import streamlit.legacy_caching.caching as legacy_caching
-    import streamlit.caching as caching
+    import streamlit.runtime.caching as caching
     from streamlit.elements.utils import clean_text
     from streamlit.proto.Spinner_pb2 import Spinner as SpinnerProto
