+++ lib/streamlit/runtime/caching/singleton_decorator.py
