+++ lib/tests/streamlit/runtime/app_session_test.py
@@ -118,8 +118,8 @@ class AppSessionTest(unittest.TestCase):
         self.assertTrue("foo" not in session._session_state)
 
     @patch("streamlit.legacy_caching.clear_cache")
-    @patch("streamlit.caching.memo.clear")
-    @patch("streamlit.caching.singleton.clear")
+    @patch("streamlit.runtime.caching.memo.clear")
+    @patch("streamlit.runtime.caching.singleton.clear")
     def test_clear_cache_all_caches(
         self, clear_singleton_cache, clear_memo_cache, clear_legacy_cache
     ):
