+++ lib/tests/streamlit/runtime/caching/hashing_test.py
@@ -14,15 +14,14 @@
 
 """st.memo/singleton hashing tests."""
 
-from dataclasses import dataclass
 import functools
 import hashlib
 import os
 import re
-
 import tempfile
 import types
 import unittest
+from dataclasses import dataclass
 from io import BytesIO
 from io import StringIO
 from unittest.mock import Mock, MagicMock
@@ -32,8 +31,8 @@ import numpy as np
 import pandas as pd
 from parameterized import parameterized
 
-from streamlit.caching.cache_errors import UnhashableTypeError
-from streamlit.caching.hashing import (
+from streamlit.runtime.caching.cache_errors import UnhashableTypeError
+from streamlit.runtime.caching.hashing import (
     _CacheFuncHasher,
     _PANDAS_ROWS_LARGE,
     _NP_SIZE_LARGE,
