+++ lib/tests/streamlit/web/cli_test.py
@@ -355,8 +355,8 @@ class CliTest(unittest.TestCase):
         self.assertEqual(0, result.exit_code)
 
     @patch("streamlit.legacy_caching.clear_cache")
-    @patch("streamlit.caching.memo.clear")
-    @patch("streamlit.caching.singleton.clear")
+    @patch("streamlit.runtime.caching.memo.clear")
+    @patch("streamlit.runtime.caching.singleton.clear")
     def test_cache_clear_all_caches(
         self, clear_singleton_cache, clear_memo_cache, clear_legacy_cache
     ):
